add: validação para preço negativo + alterar para Double

+++ fase-05/ms-produtos/src/main/java/br/com/fiap/techchallenge/produtos/model/dtos/ProdutoDTO.java
@@ -3,8 +3,6 @@ package br.com.fiap.techchallenge.produtos.model.dtos;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 
-import java.math.BigDecimal;
-
 public record ProdutoDTO(
         String id,
         @NotBlank
@@ -12,6 +10,6 @@ public record ProdutoDTO(
         @NotBlank
         String descricao,
         @DecimalMin(value = "0.0")
-        BigDecimal preco,
+        Double preco,
         Long estoque) {
 }
